fix(provider/gce): convert Internal HTTP LB impl to ts
